backend/requirements.txt
# Flask Framework
Flask==3.1.1
Flask-CORS==4.0.0
gunicorn==21.2.0

# Google Ads API - المكتبة الرسمية
google-ads==28.0.0

# Google API Client Libraries
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Google Generative AI
google-generativeai==0.8.3

# HTTP Requests
requests==2.31.0
aiohttp==3.9.1
httplib2==0.22.0

# Environment Variables
python-dotenv==1.0.0

# Data Processing
PyYAML==6.0.1
pandas>=2.2.0
numpy>=1.26.0

# Authentication & Security
passlib==1.7.4
PyJWT>=2.10.1
cryptography>=42.0.0
bcrypt>=4.1.2

# Rust Dependencies (for DLL issues)
setuptools-rust==1.8.1
wheel>=0.42.0

# Additional Security
python-jose[cryptography]==3.3.0

# Web Scraping & Parsing
beautifulsoup4==4.12.2

# Language Detection (Google's algorithm)
langdetect==1.0.9

# Data Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Console Output
rich==13.7.0

# SSL & Security
certifi>=2023.11.17

# Protocol Buffers (مطلوب لـ Google Ads API)
protobuf>=4.25.0

# gRPC (مطلوب لـ Google Ads API)
grpcio>=1.59.0
grpcio-tools>=1.59.0

# Additional Google Libraries
googleapis-common-protos>=1.63.2

# Database - Supabase (Railway compatible)
supabase==1.0.4
httpx==0.24.1
postgrest>=0.10.8,<0.11.0
# Fix for proxy argument issue
urllib3==1.26.18